editing get equipments

+++ Backend/Models/NumericalStatisticsModel.cs
@@ -1,12 +1,12 @@
 namespace Backend.Models{
     public class NumericalStatistics{
-        public int? Total_Number_Of_Clients{get; set;}
-        public int? Total_Number_Of_Coaches{get; set;}
-        public int? Total_Number_Of_Branch_Managers{get; set;}
-        public int? Total_Number_Of_Branches{get; set;}
-        public int? Total_Number_Of_Equipments{get; set;}
-        public int? Total_Number_Of_Coaches_Per_Branch{get; set;}
-        public int? Total_Number_Of_Equipments_Per_Branch{get; set;}
+        public long? Total_Number_Of_Clients{get; set;}
+        public long?Total_Number_Of_Coaches{get; set;}
+        public long? Total_Number_Of_Branch_Managers{get; set;}
+        public long? Total_Number_Of_Branches{get; set;}
+        public long? Total_Number_Of_Equipments{get; set;}
+        public long? Total_Number_Of_Coaches_Per_Branch{get; set;}
+        public long? Total_Number_Of_Equipments_Per_Branch{get; set;}
 
     }
 }
